Add robot publishing system
Instead of using a robot's code as the single source of truth, create a `PublishedRobots` table which stores different published robot versions. A robot will only be put in the matchmaking queue once the user "publishes" it. The code stored inside of `Robots` is still the code that is loaded into The Garage, but the matchmaking/public facing code is the code of the most recent robot in `PublishedRobots`. Add a new `publish` view/controller.