Better integrate publishing system into model
Break the Robot data type into two case classes: the full case class
that the db returns, and a "basic" case class that contains most of the
fields but not the code. Refactor controllers to return the basic class
in most instances, and create separate controller methods for viewing
the code.

In preparation for the opponent selection system, create a
`getRobotCode` route that returns the dev code if the user requesting it
is the owner, and the published code otherwise. Also create a
`getUserRobots` route with a similar behavior.

Split the `authAction` into two parts. The base `apply` no longer forces
a session to exist. The old behavior is replicated by `force`, which
returns a BadRequest if the user is not logged in.