Mothership commit: cleanup and seasons feature
House each battle and published robot within a "board", which can
optionally belong to a timed "season." This encourages site use because
seasons last a limited amount of time, which lowers the commitment
required to play. Also refactor a lot of the codebase: improve naming,
use case classes instead of tuples, use case classes instead of `Long`
for different ids.