Improve data validation and fix login username duplication error
Separate out validators, and perform data validation exclusively on the level of the form. Instead of transforming data, like making it all lowercase, make this into a validation, so that the data is put into the database exactly as it is entered (when it is entered correctly). Fix a bug where user username and email uniqueness wasn't checked correctly.